refactor(config): type file moved
