+++ packages/config/src/type.ts
